# community/go-engines-community/docs/swagger.yaml
basePath: /api/v4
definitions:
...
  common.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  common.PaginatedListResponse:
    properties:
      data:
        type: object
      meta:
        $ref: '#/definitions/common.PaginatedMeta'
        type: object
    type: object
  common.PaginatedMeta:
    properties:
      page:
        type: integer
      page_count:
        type: integer
      per_page:
        type: integer
      total_count:
        type: integer
    type: object
  common.ValidationErrorResponse:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
    type: object
...
  pattern.AlarmPattern:
    properties:
      entityID:
        $ref: '#/definitions/pattern.StringPattern'
        type: object
      id:
        $ref: '#/definitions/pattern.StringPattern'
        type: object
      shouldBeNil:
        type: boolean
      shouldNotBeNil:
        description: |-
          ShouldNotBeNil is a boolean indicating that the alarm should not be nil,
          and ShouldBeNil is a boolean indicating that the alarm should be nil.
          The two booleans are needed to be able to make the difference between
          the case where no alarm pattern was defined (in which case the alarm can
          be nil or not), the case where a nil alarm pattern was defined (in which
          case the alarm should be nil), and the case where a non-nil alarm
          pattern was defined (in which case the alarm should not be nil).
        type: boolean
      time:
        $ref: '#/definitions/pattern.TimePattern'
        type: object
      unexpectedFields:
        additionalProperties: true
        description: |-
          When unmarshalling a BSON document, the fields of this document that are
          not defined in this struct are added to UnexpectedFields.
        type: object
      value:
        $ref: '#/definitions/pattern.AlarmValuePattern'
        type: object
    type: object
  pattern.AlarmPatternList:
    properties:
      patterns:
        items:
          $ref: '#/definitions/pattern.AlarmPattern'
        type: array
      set:
        description: |-
          isSet is a boolean indicating whether the AlarmPatternList has been set
          explicitly or not.
        type: boolean
      valid:
        description: |-
          isValid is a boolean indicating whether the event patterns or valid or
          not.
          isValid is also false if the AlarmPatternList has not been set.
        type: boolean
    type: object
...
  pattern.AlarmValuePattern:
    properties:
      ack:
        $ref: '#/definitions/pattern.AlarmStepRefPattern'
        type: object
      activationDate:
        $ref: '#/definitions/pattern.TimeRefPattern'
        type: object
      canceled:
        $ref: '#/definitions/pattern.AlarmStepRefPattern'
        type: object
      children:
        $ref: '#/definitions/pattern.StringArrayPattern'
        type: object
      component:
        $ref: '#/definitions/pattern.StringPattern'
        type: object
      connector:
        $ref: '#/definitions/pattern.StringPattern'
        type: object
      connectorName:
        $ref: '#/definitions/pattern.StringPattern'
        type: object
      creationDate:
        $ref: '#/definitions/pattern.TimePattern'
        type: object
      displayName:
        $ref: '#/definitions/pattern.StringPattern'
        type: object
      done:
        $ref: '#/definitions/pattern.AlarmStepRefPattern'
        type: object
      extra:
        additionalProperties:
          $ref: '#/definitions/pattern.InterfacePattern'
        type: object
      hardLimit:
        $ref: '#/definitions/pattern.IntegerRefPattern'
        type: object
      initialLongOutput:
        $ref: '#/definitions/pattern.StringPattern'
        type: object
      initialOutput:
        $ref: '#/definitions/pattern.StringPattern'
        type: object
      lastEventDate:
        $ref: '#/definitions/pattern.TimePattern'
        type: object
      lastUpdateDate:
        $ref: '#/definitions/pattern.TimePattern'
        type: object
      longOutput:
        $ref: '#/definitions/pattern.StringPattern'
        type: object
      output:
        $ref: '#/definitions/pattern.StringPattern'
        type: object
      parents:
        $ref: '#/definitions/pattern.StringArrayPattern'
        type: object
      resolved:
        $ref: '#/definitions/pattern.TimeRefPattern'
        type: object
      resource:
        $ref: '#/definitions/pattern.StringPattern'
        type: object
      snooze:
        $ref: '#/definitions/pattern.AlarmStepRefPattern'
        type: object
      state:
        $ref: '#/definitions/pattern.AlarmStepRefPattern'
        type: object
      stateChangesSinceStatusUpdate:
        $ref: '#/definitions/pattern.IntegerPattern'
        type: object
      status:
        $ref: '#/definitions/pattern.AlarmStepRefPattern'
        type: object
      ticket:
        $ref: '#/definitions/pattern.AlarmTicketRefPattern'
        type: object
      totalStateChanges:
        $ref: '#/definitions/pattern.IntegerPattern'
        type: object
      unexpectedFields:
        additionalProperties: true
        description: |-
          When unmarshalling a BSON document, the fields of this document that are
          not defined in this struct are added to UnexpectedFields.
        type: object
    type: object
  pattern.BoolPattern:
    properties:
      set:
        description: Set is a boolean indicating whether the value was set or not.
        type: boolean
      value:
        description: |-
          Value contains the value of the bool. It should only be taken into
          account if Set is true.
        type: boolean
    type: object
  pattern.EntityPattern:
    properties:
      component:
        $ref: '#/definitions/pattern.StringPattern'
        type: object
      componentInfos:
        additionalProperties:
          $ref: '#/definitions/pattern.InfoPattern'
        type: object
      enabled:
        $ref: '#/definitions/pattern.BoolPattern'
        type: object
      id:
        $ref: '#/definitions/pattern.StringPattern'
        type: object
      infos:
        additionalProperties:
          $ref: '#/definitions/pattern.InfoPattern'
        type: object
      name:
        $ref: '#/definitions/pattern.StringPattern'
        type: object
      shouldBeNil:
        type: boolean
      shouldNotBeNil:
        description: |-
          ShouldNotBeNil is a boolean indicating that the entity should not be
          nil, and ShouldBeNil is a boolean indicating that the entity should be
          nil.
          The two booleans are needed to be able to make the difference between
          the case where no entity pattern was defined (in which case the entity
          can be nil or not), the case where a nil entity pattern was defined (in
          which case the entity should be nil), and the case where a non-nil
          entity pattern was defined (in which case the entity should not be nil).
        type: boolean
      type:
        $ref: '#/definitions/pattern.StringPattern'
        type: object
      unexpectedFields:
        additionalProperties: true
        description: |-
          When unmarshalling a BSON document, the fields of this document that are
          not defined in this struct are added to UnexpectedFields.
        type: object
    type: object
  pattern.EntityPatternList:
    properties:
      patterns:
        items:
          $ref: '#/definitions/pattern.EntityPattern'
        type: array
      set:
        description: |-
          Set is a boolean indicating whether the EntityPatternList has been set
          explicitly or not.
        type: boolean
      valid:
        description: |-
          Valid is a boolean indicating whether the event patterns or valid or
          not.
          Valid is also false if the EntityPatternList has not been set.
        type: boolean
    type: object
...
  pattern.InfoPattern:
    properties:
      description:
        $ref: '#/definitions/pattern.StringPattern'
        type: object
      name:
        $ref: '#/definitions/pattern.StringPattern'
        type: object
      shouldNotBeSet:
        type: boolean
      unexpectedFields:
        additionalProperties: true
        description: |-
          When unmarshalling a BSON document, the fields of this document that are
          not defined in this struct are added to UnexpectedFields.
        type: object
      value:
        $ref: '#/definitions/pattern.StringPattern'
        type: object
    type: object
...
  pattern.StringPattern:
    properties:
      equal:
        $ref: '#/definitions/types.OptionalString'
        description: |-
          If Equal is set, the value of a field has to be equal to the value of
          Equal to be matched by the pattern.
        type: object
      regexMatch:
        $ref: '#/definitions/types.OptionalRegexp'
        description: |-
          If RegexMatch is set, the value of a field has to be matched by this
          regular expression to be matched by the pattern.
        type: object
      unexpectedFields:
        additionalProperties: true
        description: |-
          When unmarshalling a BSON document, the fields of this document that are
          not defined in this struct are added to UnexpectedFields.
          UnexpectedFields should always be empty.
        type: object
    type: object
...
  pattern.TimePattern:
    properties:
      equal:
        $ref: '#/definitions/types.OptionalInt64'
        description: |-
          If Equal is set, the value of a field has to be equal to the value of
          Equal to be matched by the pattern.
        type: object
      gt:
        $ref: '#/definitions/types.OptionalInt64'
        description: |-
          If Gt is set, the value of a field has to be greater than the value
          of Gt to be matched by the pattern.
        type: object
      gte:
        $ref: '#/definitions/types.OptionalInt64'
        description: |-
          If Gte is set, the value of a field has to be greater than the value
          of Gte to be matched by the pattern.
        type: object
      lt:
        $ref: '#/definitions/types.OptionalInt64'
        description: |-
          If Lt is set, the value of a field has to be greater than the value
          of Lt to be matched by the pattern.
        type: object
      lte:
        $ref: '#/definitions/types.OptionalInt64'
        description: |-
          If Lte is set, the value of a field has to be greater than the value
          of Lte to be matched by the pattern.
        type: object
      unexpectedFields:
        additionalProperties: true
        description: |-
          When unmarshalling a BSON document, the fields of this document that are
          not defined in this struct are added to UnexpectedFields.
          UnexpectedFields should always be empty.
        type: object
    type: object
...
  playlist.EditRequest:
    properties:
      enabled:
        type: boolean
      fullscreen:
        type: boolean
      interval:
        $ref: '#/definitions/types.DurationWithUnit'
        type: object
      name:
        type: string
      tabs_list:
        items:
          type: string
        type: array
    required:
    - enabled
    - fullscreen
    - interval
    - name
    - tabs_list
    type: object
  playlist.Playlist:
    properties:
      _id:
        type: string
      author:
        type: string
      created:
        type: integer
      enabled:
        type: boolean
      fullscreen:
        type: boolean
      interval:
        $ref: '#/definitions/types.DurationWithUnit'
        type: object
      name:
        type: string
      tabs_list:
        items:
          type: string
        type: array
      updated:
        type: integer
    type: object
...
  scenario.Action:
    properties:
      alarm_patterns:
        $ref: '#/definitions/pattern.AlarmPatternList'
        type: object
      drop_scenario_if_not_matched:
        type: boolean
      emit_trigger:
        type: boolean
      entity_patterns:
        $ref: '#/definitions/pattern.EntityPatternList'
        type: object
      parameters:
        additionalProperties: true
        type: object
      type:
        type: string
    type: object
  scenario.ActionRequest:
    properties:
      alarm_patterns:
        $ref: '#/definitions/pattern.AlarmPatternList'
        type: object
      drop_scenario_if_not_matched:
        type: boolean
      emit_trigger:
        type: boolean
      entity_patterns:
        $ref: '#/definitions/pattern.EntityPatternList'
        type: object
      parameters:
        type: object
      type:
        type: string
    required:
    - drop_scenario_if_not_matched
    - emit_trigger
    - type
    type: object
  scenario.EditRequest:
    properties:
      actions:
        items:
          $ref: '#/definitions/scenario.ActionRequest'
        type: array
      author:
        type: string
      delay:
        $ref: '#/definitions/types.DurationWithUnit'
        type: object
      disable_during_periods:
        items:
          type: string
        type: array
      enabled:
        type: boolean
      name:
        type: string
      priority:
        type: integer
      triggers:
        items:
          type: string
        type: array
    required:
    - actions
    - author
    - enabled
    - name
    - priority
    - triggers
    type: object
  scenario.Scenario:
    properties:
      _id:
        type: string
      actions:
        items:
          $ref: '#/definitions/scenario.Action'
        type: array
      author:
        type: string
      created:
        type: integer
      delay:
        $ref: '#/definitions/types.DurationWithUnit'
        type: object
      disable_during_periods:
        items:
          type: string
        type: array
      enabled:
        type: boolean
      name:
        type: string
      priority:
        type: integer
      triggers:
        items:
          type: string
        type: array
      updated:
        type: integer
    type: object
...
  types.DurationWithUnit:
    properties:
      seconds:
        type: integer
      unit:
        type: string
    required:
    - seconds
    - unit
    type: object
...
  types.OptionalInt64:
    properties:
      set:
        description: Set is a boolean indicating whether the value was set or not.
        type: boolean
      value:
        description: |-
          Value contains the value of the int64. It should only be taken into
          account if Set is true.
        type: integer
    type: object
  types.OptionalRegexp:
    properties:
      set:
        description: Set is a boolean indicating whether the value was set or not.
        type: boolean
      value:
        $ref: '#/definitions/utils.RegexExpression'
        description: |-
          Value contains the value of the regular expression. It should only be
          taken into account if Set is true.
        type: object
    type: object
  types.OptionalString:
    properties:
      set:
        description: Set is a boolean indicating whether the value was set or not.
        type: boolean
      value:
        description: |-
          Value contains the value of the string. It should only be taken into
          account if Set is true.
        type: string
    type: object
...
info:
  contact: {}
  description: This is a Canopsis server.
  license: {}
  title: Canopsis API
  version: "4.0"
paths:
...
  /playlists:
    get:
      consumes:
      - application/json
      description: Get paginated list of playlist
      operationId: playlists-find-all
      parameters:
      - description: current page
        in: query
        name: page
        required: true
        type: integer
      - description: items per page
        in: query
        name: limit
        required: true
        type: integer
      - description: search query
        in: query
        name: search
        type: string
      - description: sort query
        in: query
        name: sort
        type: string
      - description: sort query
        in: query
        name: sort_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.PaginatedListResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/playlist.Playlist'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ValidationErrorResponse'
      security:
      - ApiKeyAuth: []
      - BasicAuth: []
      summary: Find all playlist
      tags:
      - playlists
    post:
      consumes:
      - application/json
      description: Create playlist
      operationId: playlists-create
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/playlist.EditRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/playlist.Playlist'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ValidationErrorResponse'
      security:
      - ApiKeyAuth: []
      - BasicAuth: []
      summary: Create playlist
      tags:
      - playlists
  /playlists/{id}:
    delete:
      description: Delete playlist by id
      operationId: playlists-delete-by-id
      parameters:
      - description: playlist id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204": {}
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BasicAuth: []
      summary: Delete playlist by id
      tags:
      - playlists
    get:
      description: Get playlist by id
      operationId: playlists-get-by-id
      parameters:
      - description: playlist id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/playlist.Playlist'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BasicAuth: []
      summary: Get playlist by id
      tags:
      - playlists
    put:
      consumes:
      - application/json
      description: Update playlist by id
      operationId: playlists-update-by-id
      parameters:
      - description: playlist id
        in: path
        name: id
        required: true
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/playlist.Playlist'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/playlist.Playlist'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ValidationErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BasicAuth: []
      summary: Update playlist by id
      tags:
      - playlists
...
  /scenarios:
    get:
      consumes:
      - application/json
      description: Get paginated list of scenarios
      operationId: scenarios-find-all
      parameters:
      - description: current page
        in: query
        name: page
        required: true
        type: integer
      - description: items per page
        in: query
        name: limit
        required: true
        type: integer
      - description: search query
        in: query
        name: search
        type: string
      - description: sort query
        in: query
        name: sort
        type: string
      - description: sort query
        in: query
        name: sort_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.PaginatedListResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/scenario.Scenario'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ValidationErrorResponse'
      security:
      - ApiKeyAuth: []
      - BasicAuth: []
      summary: Find scenarios
      tags:
      - scenarios
    post:
      consumes:
      - application/json
      description: Create scenario
      operationId: scenarios-create
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/scenario.EditRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/scenario.Scenario'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ValidationErrorResponse'
      security:
      - ApiKeyAuth: []
      - BasicAuth: []
      summary: Create scenario
      tags:
      - scenarios
  /scenarios/{id}:
    delete:
      description: Delete scenario by id
      operationId: scenarios-delete-by-id
      parameters:
      - description: scenario id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204": {}
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BasicAuth: []
      summary: Delete scenario by id
      tags:
      - scenarios
    get:
      description: Get scenario by id
      operationId: scenarios-get-by-id
      parameters:
      - description: scenario id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/scenario.Scenario'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BasicAuth: []
      summary: Get scenario by id
      tags:
      - scenarios
    put:
      consumes:
      - application/json
      description: Update scenario by id
      operationId: scenarios-update-by-id
      parameters:
      - description: scenario id
        in: path
        name: id
        required: true
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/scenario.EditRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/scenario.Scenario'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ValidationErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BasicAuth: []
      summary: Update scenario by id
      tags:
      - scenarios
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-canopsis-authkey
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"